+++ packages/core/src/ElementOption.ts
@@ -28,6 +28,9 @@ export interface RunArguments {
 	verbose?: boolean
 	browser?: BrowserType
 	export?: boolean
+	executablePath?: string
+	downloadsPath?: string
+	showScreenshot?: boolean
 }
 
 export interface ElementRunArguments {
@@ -57,6 +60,8 @@ export interface ElementRunArguments {
 	mu: boolean
 	runArgs: RunArguments
 	testSettings?: TestSettings
+	executablePath?: string
+	downloadsPath?: string
 }
 
 export interface ElementOptions {
@@ -78,6 +83,8 @@ export interface ElementOptions {
 	failStatusCode: number
 	browser: BrowserType
 	export?: boolean
+	executablePath?: string
+	downloadsPath?: string
 	showScreenshot?: boolean
 }
 
@@ -212,6 +219,8 @@ export function normalizeElementOptions(
 		failStatusCode: args['fail-status-code'],
 		browser: args.browser,
 		export: args.export,
+		executablePath: args.executablePath ?? '',
+		downloadsPath: args.downloadsPath ?? '',
 		showScreenshot: args.showScreenshot,
 	}
 
@@ -240,8 +249,10 @@ export function normalizeElementOptions(
 		opts.headless = runArgs.headless ?? opts.headless
 		opts.devtools = runArgs.devtools ?? opts.devtools
 		opts.sandbox = runArgs.sandbox ?? opts.sandbox
+		opts.export = runArgs.export ?? opts.export
 		opts.browser = runArgs.browser ?? opts.browser
-		opts.export = !!runArgs.export ?? opts.export
+		opts.executablePath = runArgs.executablePath ?? opts.executablePath
+		opts.downloadsPath = runArgs.downloadsPath ?? opts.downloadsPath
 	}
 
 	opts.testSettingOverrides = setupDelayOverrides(args, opts.testSettingOverrides)