+++ packages/core/src/runtime/Settings.ts
@@ -213,7 +213,7 @@ export interface TestSettings {
 	/**
 	 * Specifies options to launch
 	 */
-	browserLaunchOption?: BrowserLaunchOptions
+	browserLaunchOptions?: BrowserLaunchOptions
 
 	/**
 
@@ -286,7 +286,7 @@ export const DEFAULT_SETTINGS: ConcreteTestSettings = {
 	device: null,
 	ignoreHTTPSError: false,
 	browser: 'chromium',
-	browserLaunchOption: {},
+	browserLaunchOptions: {},
 	blockedDomains: [],
 	incognito: false,
 	name: 'Element Test',